requirements.txt
streamlit
supabase
pandas
google-api-python-client
google-auth-oauthlib
google-generativeai
Pillow
python-dotenv
